notion/requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
watchdog>=3.0.0
psutil>=5.9.0
schedule>=1.2.0


